+++ archiver/src/main/java/com/netflix/conductor/archiver/job/EventExecsJob.java
@@ -31,7 +31,7 @@ public class EventExecsJob extends AbstractJob {
 			logger.info("Deleting records earlier than " + endTime + ", batch size = " + batchSize);
 
 			int deleted = 0;
-			List<Integer> ids = fetchIds(QUERY, endTime, batchSize);
+			List<Long> ids = fetchIds(QUERY, endTime, batchSize);
 			while (isNotEmpty(ids)) {
 				deleted += deleteByIds("event_execution", ids);
 				logger.debug("EventExecs job deleted " + deleted);
